+++ packages/reactabular-table/src/index.js
@@ -5,4 +5,5 @@ export { default as BodyRow } from './body-row';
 export { default as evaluateFormatters } from './evaluate-formatters';
 export { default as evaluateTransforms } from './evaluate-transforms';
 export { default as mergeProps } from './merge-props';
+export { default as columnsAreEqual } from './columns-are-equal';