+++ packages/reactabular-virtualized/src/body-row.js
@@ -1,6 +1,6 @@
 import { isEqual } from 'lodash';
 import React from 'react';
-import { columnsAreEqual } from 'reactabular-utils';
+import { columnsAreEqual } from 'reactabular-table';
 import { bodyRowContextTypes, bodyRowTypes } from './types';
 
 class BodyRow extends React.Component {