+++ packages/reactabular/CHANGELOG.md
@@ -70,7 +70,7 @@ render() {
   * Feature - Pass whole column through header/body for extra parameters.
   * Feature - Support `onRow` at `Table.Header`.
   * Feature - Allow `Table.Header` to accept `headerRows` (an array of column definitions) to override default columns. See below.
-  * Feature - Move `utils.evaluateFormatters`, `utils.evaluateTransforms`, and `utils.mergeProps` to `reactabular-table`. Those aren't used elsewhere so that makes a lot of sense.
+  * Feature - Move `utils.evaluateFormatters`, `utils.evaluateTransforms`, `utils.mergeProps`, `utils.columnsAreEqual` to `reactabular-table`.
   * Bug fix - Skip functions at `BodyRow` `shouldComponentUpdate`.
   * Breaking - Generalize `format: <fn>` as `formatters: [<fn>]`. The formatters are applied recursively from left to right: `[f1, f2, f3] => f1(f2(f3(value, extra)))`. This allows composition.
   * Breaking - Extract nested column logic. Now you will have to resolve nested columns before passing them to the table. The advantage of doing this is that now all logic (search/sorting/etc.) works with nested tables. Basic idea: